[ch37171] Create header component in `etools-unicef` to be used in all FE apps in order to standardize layout and be mobile-friendly
