+++ package.json
@@ -21,7 +21,8 @@
     "rd:init": "relative-deps init",
     "rd:add": "relative-deps add",
     "rd:reload": "relative-deps",
-    "rd:watch": "chokidar \"../../etools-polymer/**/*.js\" \"../../etools-polymer/**/**/*.js\" -c 'npm run rd:reload' --silent"
+    "rd:watch": "chokidar \"../../etools-polymer/**/*.js\" \"../../etools-polymer/**/**/*.js\" -c 'npm run rd:reload' --silent",
+    "prepare": "relative-deps"
   },
   "dependencies": {
     "@mapbox/leaflet-omnivore": "^0.3.4",
@@ -33,8 +34,8 @@
     "@unicef-polymer/etools-modules-common": "^5.0.0-rc.50",
     "@unicef-polymer/etools-piwik-analytics": "^4.0.1-rc.3",
     "@unicef-polymer/etools-types": "^1.11.17",
-    "@unicef-polymer/etools-unicef": "^1.0.0-rc.86",
-    "@unicef-polymer/etools-utils": "^2.0.0-rc.7",
+    "@unicef-polymer/etools-unicef": "^1.0.0-rc.89",
+    "@unicef-polymer/etools-utils": "^2.0.0-rc.9",
     "dayjs": "^1.11.10",
     "dexie": "^3.2.4",
     "esri-leaflet": "^3.0.8",
