--- src_ts/components/app-shell/header/page-header-styles.ts
@@ -1,103 +1,0 @@
-import {css, CSSResult} from 'lit';
-
-// language=CSS
-export const pageHeaderStyles: CSSResult = css`
-  app-toolbar {
-    padding: 0;
-    flex-wrap: wrap;
-    height: 100%;
-    justify-content: space-between;
-  }
-
-  .titlebar {
-    color: var(--header-color);
-  }
-
-  #menuButton {
-    display: flex;
-    color: var(--header-color);
-  }
-
-  support-btn {
-    color: var(--header-color);
-    margin-left: auto;
-  }
-
-  .titlebar {
-    display: flex;
-    font-size: var(--etools-font-size-28, 28px);
-    font-weight: 300;
-  }
-
-  .titlebar img {
-    width: 34px;
-    margin: 0 8px 0 24px;
-  }
-
-  .content-align {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-  }
-
-  #app-logo {
-    height: 32px;
-    width: auto;
-  }
-
-  .envWarning {
-    color: var(--header-color);
-    font-weight: 700;
-    font-size: var(--etools-font-size-18, 18px);
-    line-height: 20px;
-  }
-  @media (max-width: 1024px) {
-    .envWarning {
-      display: none;
-    }
-    .envLong {
-      display: none;
-    }
-    .titlebar img {
-      margin: 0 8px 0 12px;
-    }
-    etools-profile-dropdown {
-      width: 40px;
-      margin-inline-end: 10px;
-      margin-inline-start: -5px;
-    }
-  }
-  @media (max-width: 576px) {
-    #app-logo {
-      display: none;
-    }
-    .titlebar img {
-      margin: 0 8px 0 4px;
-    }
-
-    .refresh-button {
-      margin-right: 0;
-    }
-    .envWarning {
-      font-size: 10px;
-      line-height: 12px;
-      white-space: nowrap;
-      margin-left: 2px;
-    }
-    .dropdown {
-      order: 1;
-    }
-  }
-  @media (min-width: 850px) {
-    #menuButton {
-      display: none;
-    }
-  }
-  @media (max-width: 490px) {
-    app-toolbar {
-      height: auto;
-      padding-inline-end: 0px;
-      margin: 0 !important;
-    }
-  }
-`;
